tests: removed 'retargetable' framework
I am not 100% sure what that is but it seems like an experiment that
never gave fruits. Since the introduction of the code, we had huge
changes in how we test api, we got tempest plugins, we steadily move
from api testing in unit tests, we have fullstack, etc.

I assume if I can't understand what's use of that, then it's probably
irrelevant code, and so it's better to clean it from the tree.

Change-Id: I0b1be0c761da75885981952a4f5ab5a43bc09525