Switched from q-dns to neutron-dns service name
The q-* service names make devstack use old lib/neutron-legacy library
to configure neutron services. We plan to swith to the new lib/neutron
library, and it's a requirement that no services with q-* names are
present in ENABLED_SERVICES to make it happen.

The neutron devstack plugin handles both names identically, so there
should be no functional change after the patch is merged.

Change-Id: Ic5100cd2f12b729149350e3ccb0aba7844b18d2c