Do not assume all protocols in neutron-lib are usable for SG
Tempest is branchless and thus supposed to work against older
deployments.  Also, it's supposed to work against other
implementations, which might not support the same set as
the reference implementation.  Ideally SG can have a way
to discover the set of usable protocols.  But for now,
let's make this conservative a bit.  It might still be too
aggressive because the set of protocols here was taken from
Ocata. (neutron-lib 1.1.0)

Closes-Bug: #1742385
Change-Id: Id67037252aafbacbe43f6af786f30a687321952d