Make test_create_port_when_quotas_is_full more generic
This patch is making the test test_create_port_when_quotas_is_full more
generic to work with different ML2 drivers.

The test is being changed to set the quotas to the number of existing
ports + 1, instead of having it hardcoded to 1. The reason for that is
because the ML2/OVN driver does create a port to be used by the metadata
service upon creating a new network. That would then already fulfil the
ports quota for that tenant.

This approach was suggested at
http://lists.openstack.org/pipermail/openstack-dev/2018-April/129084.html

Closes-Bug: #1761119
Change-Id: I86fe55eb48b9964b33d4b5174a248676129abfb1