Allow to connect to SSH server using an intermediate SSH server
For debugging purposes it could be handy to execute tests
on your workstation and connect test instances via SSH passing
throw an intermediate SSH server.

This allow to configure in tempest.conf an intermediate
SSH client connection to be used from tests to create SSH
connections to VMs.

Example of configuration in tempest.conf:

  [neutron_plugin_options]
  ssh_proxy_jump_host = some.ssh.server
  ssh_proxy_jump_username = root
  # ssh_proxy_jump_password = # better using keys
  proxy_jump_keyfile = ~/.ssh/id_rsa
  proxy_jump_port = 22

Change-Id: Icae73c2cddbdcd8da2b4cdb07a7027791642c6a8