Allow creating subnet without a gateway
The ability to create a subnet without a gateway was broken
in commit 0ddc93b1b19922d08bedf331b57c363535bb357e.  Change
the create_subnet() method back to using '' as the default
value for the gateway so that callers can pass-in None as
before.  This allows the trunk test to revert to its previous
behavior of passing None.

Change-Id: Ib010f277f4c9ad8a708d16debdcdf8d2ceccd051
Closes-Bug: #1769609