Allow skipping subnet CIDRs reservation when creating subnet.
This implements a special parameter to be passed to
create_subnet method to specify subnet CIDR will not be
reserved after its assigment to a new subnet.

It also reserve subnet CIDR only after subnet has been
sucessefully created.

Reserved CIDRs are still not released when subnets are
deleted. This is is left for future improvements.

Change-Id: I421dd28999086ee0af9246121f4f033291e341b4
Related-Bug: #1766702