Add guest image customization to DevStack plugin.
CI jobs for this plugin use DevStack to install and configure
OpenStack and to upload guest images to be used for tests.

This change allow to customize guest images just before
being sent to image service with the purpose of installing
new packages (like socat), configuring N network devices, etc.

This change has been created because for testing multicast
socat is required, but it was designed with the idea
of being used to allow further guest images customizations.

Change-Id: I88491dbb65031fe3743d1c3d27c38a57b5511794