Skip filter validation tests
There is an on-going effort on adding support for strict filter
validation [1]. During the transition period when neutron is
migrating to strict filter validation, the filter validation
tests might fail. According to the tempest's suggestion [2],
this patch temporarily skips those tests, which is necessary to
make the CI happy to land the server side patches.
These tests will be modified and re-enabled after the implemention
is finished in server side.

[1] https://review.openstack.org/#/c/574907/
[2] https://docs.openstack.org/tempest/latest/HACKING.html
    #bug-fix-on-core-project-needing-tempest-changes

Related-Bug: #1749820
Change-Id: I057cd917628c77dd20c0ff7747936c3fec7b4844