Add convenience parameters to create_network.
It adds convenience parameters to crete_network method:
- external -> shortcut to router:external
- provider_network_type -> shorcut to provider:network_type
- provider_physical_network -> shortcut to provider:phyiscal_network
- provider_segmentation_id -> shortcut to provider:segmentation_id

When client is not provider and admin_client is attribute is not None
(for example when using BaseAdminNetworkTest base class) and using any
of the convenience parameters it silently uses admin_client. If the
network is not shared then it uses the same project_id as regular
client.

It also add delete_network to help deleting networks
created with other clients than cls.client.

When admin_client attribute is not None (for exemble when using
BaseAdminNetworkTest base class), for convenience it
silently uses admin_client when required, while using the same
project ID as regular client when network is not shared.

Change-Id: I1ef57c518c137f1dbbd3e345b48afe9a35b10904