Assert extensions match in config file and server
If an API extension is enabled in tempest config file, we want
to ensure the extension is present in neutron server.
In before, we occasionally ran into issues that the list of
extensions is mismatch, which leads to failure of test cases.
This patch adds a test case to check that.

Change-Id: I4c59f407c30020e3865e96e67a1b61361e699014