Fix bash cmd used in scenario trunk tests
In scenario trunk test test_subport_connectivity there
is bash command used to check name of base interface on
spawned vm (e.g. ens3 or eth0), configure vlan interface
with vlan_id used on trunk port (e.g. ens3.10),
make this vlan device up and run dhclient on this interface
if it's not running yet.

This command was broken and that cause failure of this test.

Change-Id: I4c0207f79cd6df2594f976b9509697209011edf2
Closes-Bug: #1766701