Add release notes job to zuul template
We are not running the release notes job, so are unable to
see the built release notes when they are included in a patch.
Add it as in the other neutron repos.

Also added a similar README.rst as the neutron repo to give
the basic info on release note generation, and fixed up
various issues in the configuration description.

Change-Id: I98ffed3ed77a16bba59cae69a73cb699ddb175fb