Create wrapper tool for both local and remote shell command execution.
Create base function called execute that takes a command and execute
it using either subprocess module or an SSH client to allow
implementing commands wrapper that can be executed either locally or
remotelly.

It also try to handle timeouts and produce good error capable of
reporting whats written to stdout and stderr.

Change-Id: I1a30b82338f44a4182722973e7ad3da2872295fd