Fix creating subnet without gateway
If create_subnet() method is called with gateway=None explicity,
subnet should be created without gateway_ip specified.
To achieve that "gateway_ip=null" should be passed in json in
request's body to neutron server.
This was missing, so neutron-server allocated gateway_ip automatically.
Now gateway for such network will not be set as is expected.

Closes-Bug: #1769609

Change-Id: Ia9f0646a3cf371f82f2aa2dc22837249531d1ff5