Add additional config option "default_image_is_advanced"
In patch [1] we added new config option to configure "advanced" image
which we are using in some tests which require it and can't run using
default Cirros image.
Unfortunatelly this made life of some downstream users harder as they
have CI systems which don't use devstack and don't have easy way to
load another "advanced" image and use it for those tests.

So now this patch adds one ore config option "default_image_is_advanced"
which has boolean value and defaults to False.
If this is set to True, tests which require advanced image will not be
skipped and default image will be used for those tests.

[1] https://review.openstack.org/#/c/609762/16/neutron_tempest_plugin/config.py

Co-Authored-By: Slawek Kaplonski <skaplons@redhat.com>

Change-Id: Ic78ee64a9350863ff5bd4b4baa4b20770e1b9d47