Migrate neutron-fwaas tests to neutron-tempest-plugin
As discussed in the neutron_c1 meeting [1] the QA team would like to
move the tempest tests for the stadium projects from their repos to
repos specific to being tempest plugins.  This is the first part of a
two stage move, by copying over the tempest tests to the
neutron-tempest-plugin repo [2] rather than spawning new repos to be
separate.

[1] http://eavesdrop.openstack.org/meetings/neutron_ci/2019/neutron_ci.2019-03-12-16.01.log.html#l-94
[2] https://etherpad.openstack.org/p/neutron_stadium_move_to_tempest_plugin_repo

Needed-By: https://review.opendev.org/643668
Depends-On: https://review.opendev.org/660483
Change-Id: I979edd26264ae5f9ceab2da350bc99c40145ec40