Don't customize image before upload it to Glance by default
Commit [1] provided new possibility to customize e.g. Ubuntu
image and e.g. install some additional packages just before
upload it to Glance in Devstack's stack.sh script.
It was done to be able to install e.g. socat package to test
multicast.
But until now we don't have tests which would use this and
this customization causes sometimes job failures when
there is package and index mismatch.

So there is no need to customize image always by default.
If we will need to use this customization, we should use
OpenStack's infra mirrors of Ubuntu repositories and then
enable this option for job where its needed.

[1] https://review.openstack.org/#/c/573702/

Change-Id: Ie84c0ae56dbd42cd522c4258903b7eafb435004f
Closes-Bug: #1821384