Change legacy security group rule check
A neutron change, https://review.opendev.org/#/c/453346/
is standardizing the protocol name for IPv6 ICMP in security
group rules to be 'ipv6-icmp', even if 'icmp' or 'icmpv6'
was passed during creation.  Change the API test to check
against a list of possible values so it covers both old and
new behaviors.

Change-Id: I0ca8d743ca56f7d67ef8c1ae45ca518bd6e6dc35
Partial-Bug: #1582500