Remove networking-bgpvpn job from template
Job neutron-tempest-plugin-bgpvpn-bagpipe was by mistake added
to the neutron-tempest-plugin-jobs template in [1] and because
of that it was run in check and gate queue for every Neutron
patch.
This isn't necessary so this patch moves this job out of the template
and adds it directly to check and gate queue in neutron-tempest-plugin
repo.

[1] https://review.openstack.org/652991

Change-Id: Iec9dc16b2e37d4f93dd15db314bc3286779489dd