Ensure DNS configuration from conf
For internal DNS the DNS configuration including the search
path should be derived from configuration rather than from
the dns_domain attribute of the network, which is used for
external DNS integration.

Update test to ensure that configuration based dns domain
is present in the server search path and is used to validate
resolution of the hostname of the server.

This commit also temporaily skips the internal DNS test whilst
the associated Neutron reviews are landed; it will be re-enabled
afterwards.

Change-Id: I145144c042b100f7e12a02a8ac7e0fbbe41e984d
Related-Bug: 1826419
Required-By: I5ff03b5ad8af432a9f7919ef953d7d8c434b93bd