Change order of creating vms and plug routers in scenario test
In scenario tests in module
test_connectivity there was an issue that first 2 vms
were created and after that subnets were plugged into
router. That caused race condition between spawning vms
and cloud-init script during boot process and configuring
metadata service in routers. Because of that often
instance was booted without SSH key configured properly
thus there was no possibility to ssh to this VM and test
was failing.

As we don't have any way to ensure that metadata is already
configured inside router, this patch just change order of
operations that subnets are first plugged into router and
than VMs are created. Thanks to this change it should be
at least much more reliable and test should be working better.

Change-Id: Ieca8567965789f8d7763a77cecc82059c30b5ced
Related-Bug: #1830763