Add new job neutron-tempest-plugin-scenario-openvswitch-iptables_hybrid
This new scenario job is single node job which will run
neutron-tempest-plugin scenario tests with openvswitch agent and
iptables_hybrid firewall driver.
Since in Rocky we switched default fw driver in devstack to be
"openvswitch" we are didn't have any job which would run
neutron-tempest-plugin tests on openvswitch agent with iptables_hybrid
driver. And it's necessary to have such job.

This patch don't add such job for Queens release because in Queens
iptables_hybrid was default driver so
"neutron-tempest-plugin-scenario-openvswitch" job will cover this
case for this release.

Change-Id: I37c43e0fbb434a53f1b971895dd3d6e9ab7b617f