Make one agent untouchable in test_agent_management
In test_agent_management some tests depend on one agent to be unmodified
which was not guaranteed as some tests updated that agent.
The update happened because the update test assumed that neutron will
return list of agents in different order.
The fix is only to select always different agent for update than the one
used for other tests.

Change-Id: I7af97ac0ee396d2eb132459d29e938bd5f3de9aa
Closes-Bug: #1845330