Add test to attach and detach policies with regular client
The test validates that a regular client is allowed to attach\detach
QoS policies only to\from networks with policies that were configured
as "Shared".

This test will not be running automatically because of the creation
of a "Shared" resource during the test.
Setting the create_shared_resources option as 'True' in the config.py
file is needed.

Change-Id: Iec3fda32cbce00a677cb30dac3c234c99d7c27fd