Fix intermittent port_forwarding test failures
This test was failing from time to time on checking UDP port
forwarding.
The reason for that was that sometimes (for unknown for me reason)
nc server spawned on guest vm wasn't run properly. Due to that there was
no response from this nc server and test was failing.

This patch changes that to use wait_until_true() helper and wait
60 seconds to received msg. Every time it tries to receive message,
first it tries also to spawn nc server on guest vm.

This patch also reverts commit
526f53722fc51eb1f9bdac957495518cbeae5e10.
as port_forwarding test don't need to be marked as unstable anymore.

Closes-Bug: #1850800

Change-Id: I95d5ba02b2851a07e6e1a8a2bedfdce7d20f32fc