Fix how netcat is called when started as server
Netcat started as server was started with "-e" option which,
according to documentation in [1] should be used to pass filename
of script which will be executed after connect.
In scenario tests we are not using any additional script,
but we are passing simple shell command which should be executed
("echo $msg"). That should be passed with "-c" option instead of
"-e" and this patch is changing exactly that.

[1] https://manpages.debian.org/testing/netcat-traditional/nc.1.en.html

Change-Id: Ib0a4e8bb6baf596f441539c6fb93ab2ebc4f98af