Fix neutron-tempest-plugin tag for queens jobs
neutron-tempest-plugin tag 0.3.0 is for stable/queen which is
compatible with neutron-lib==1.13.0[1].

We are pinning Tempest and upper constraint used in tempest venv
for stable/queens which pickup neutron-lib as 1.13.0 from stable/queens
upper constraint.

Currently used tag of neutron-tempest-plugin 0.7.0 for queens job
is not compatible with stable/queens neutron-lib version. we can see
the failure[2]

Also skip the dns test for queens job which fixed went in 0.4.0 tag
(https://bugs.launchpad.net/neutron/+bug/1826419) and 0.4.0 tag is not
compatible with stable/queens constraint.


Related-Bug: 1859988

[1]https://github.com/openstack/requirements/blob/f7b45b47bfc5f9b635eb5fa6613482c8c253b091/upper-constraints.txt#L58

[2]
- https://6631124a20df425843c7-d8780b91d5409425c6866f467a1249cf.ssl.cf5.rackcdn.com/703680/3/check/neutron-tempest-plugin-api-queens/bba5f0e/job-output.txt

2020-01-23 17:36:50.747528 | controller |   File "/opt/stack/neutron-tempest-plugin/neutron_tempest_plugin/bgpvpn/scenario/test_bgpvpn_basic.py", line 20, in <module>
2020-01-23 17:36:50.747537 | controller |     from neutron_lib.utils import test
2020-01-23 17:36:50.747545 | controller | ImportError: cannot import name test


Change-Id: Icc4a91c333ac746183cb423f5e6c59ee0fbdfe2f