Adding pattern to check_remote_connectivity function
Ping messages sent by this function will include a pattern of bytes
that are repeated until message size is completed.
Input format required is a string with hex digits.

This is useful when these messages are captured, in order to identify
and differentiate clearly messages from different tests.
It can be used to validate that traffic routing is not affected by
packet data as well.

Closes-Bug: #1861397

Change-Id: Ib94518597bdf3d9f3049643d3242db632769de6b