Enhance the test_multicast_between_vms_on_same_network test
This patch is enhancing the test_multicast_between_vms_on_same_network
test to fit the different scenarios that can be encountered when using
multicast:

1) When IGMP snooping is enabled and the multicast group address *is not*
   in the 224.0.0.X range [0], asserts that the multicast traffic *is not*
   flooded.

2) When IGMP snooping is not enabled, asserts that the multicast traffic
   is flooded and the unregistered VM gets it.

3) When IGMP snooping is enabled and the multicast group addres *is* in
   the 224.0.0.X range [0], asserts that the multicast traffic *is*
   flooded.

In order to make those assertions, a new VM is being launched as part of
the test running tcpdump to verify whether the traffic is reaching it or
not.

A new configuration option called "is_igmp_snooping_enabled" has been
added.

[0] https://tools.ietf.org/html/rfc4541 (See section 2.1.2)

Change-Id: I8af041925119463c7199238988f0133e8d993a8f
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>