Modify rocky jobs to run with correct py version
OpenStak ussuri release is python3-only which means
stable branhces jobs running on neutron-tempest-plugin
master gate should run with py3. neutron-tempest-plugin
will drop the py2 support soon so it will not be possible
to test the stable branch with py2.

Rocky jobs are modified to run on py2 on stable/rocky gate
and on py3 for master gate.

Change-Id: Ic04c9d1d2562398bd3640bfd064971bc82235502