Set higher timeout for tests in scenario jobs
It seems that in case of scenario tests which uses advanced image, when
tests are run on slow node, it may happen that around 20 minutes test
timeout will be not enough to perform all cleanup after test and it will
fail due to that.
So lets give more time for such tests to make sure all will be cleaned
properly.

Change-Id: Ic2177dc90a6fc8db4b50f05649ad794f6c3d5a23
Closes-Bug: #1872693