Fix test_multicast_between_vms_on_same_network
The test_multicast_between_vms_on_same_network had a few errors, one
related to the "mcast_address" being an IPAddress object and not a
string. This has been fixed by converting the object to its string
representation before trying to do a startswith() operation on it.

The second problem was related to the execution of the
unregistered_traffic_receiver.sh script. The script was created by
the default SSH user in its home directory but, when SSH'ing the VM to
execute it we had the become_root parameter set to True, which caused
the command to look for the script in the /root directory. This has been
fixed by creating and executing the script from the /tmp directory cos
it's userless.

Change-Id: I4046aaf0e7f4171e6ce5751b79b84c943ccf56cd
Closes-Bug: #1873798
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>