Add test_ping_global_ip_from_vm_with_fip
We need such a test in order to be able to catch issues like [1].
The test creates network and server, then launches a VM with a FIP.
Then connects to the FIP and pings a global ip address.
The test requires global_ip_address variable in tempest.conf.
By default the value is not set and the test will be skipped.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1834433

Change-Id: Ibb3db814ed9b42f28ced8378368a3d5c2e43d7b3