Preserve floating IP address in FloatingIpMultipleRoutersTest
In "test_reuse_ip_address_with_other_fip_on_other_router", one of the
floating IP addresses is deleted and created again using the same IP
address. As reported in the bug, the time spent between the deletion
and the creation should be reduced. That will prevent any other test
to, in parallel, create a floating IP address reusing the same IP
address of the test.

Change-Id: Ia5694208934c37efd68df0fd33eaac4c7e879c07
Closes-Bug: #1880976