Remove propagate_uplink_status API related tests
Due to the related bug we can't have test
test_create_port_without_propagate_uplink_status in
neutron-tempest-plugin repo anymore as this repo is branchless and we
run tests from it against neutron master branch and against stable
branches. Now default value of the "propagate_uplink_status" attribute
was changed in master branch but not in stable branches so this test
would fail always.

Because of that those tests are now moved to be in fullstack - see [1]
for details and we don't need them in neutron-tempest-plugin anymore so
this patch is removing them from neutron-tempest-plugin.

[1] https://review.opendev.org/#/c/745521/

Change-Id: I79581da5b2cb3df6fa9407f1f06509513ac7f96d
Closes-Bug: #1890842