Migrate CI jobs to Ubuntu Focal
Jobs for master branch are moved to be run on Ubuntu Focal.
All jobs for Stein, Train and Ussuri will be still run on Ubuntu
Bionic.

We also need to switch to legacy ebtables implementation in the
linuxbridge job because ebtables-nft implementation don't
supports syntax for source and destination ipv4 address in arp
tables. Please check bug [1] for more details.

Additionally scenario tests for port forwarding are now marked as
unstable as we have some problem with port forwarding on Ubuntu Focal.
See [2] for details.

Also test test_floating_ip_update is now marked as unstable as it is
failing pretty often on Ubuntu Focal. See [3] for details.

This patch also changes ovn hash used to be installed on the nodes
in the ovn scenario job with Ussuri release as this job is still run
on Ubuntu Bionic and we need to bump this hash there.

This patch additionally switches neutron-tempest-plugin-bgpvpn-bagpipe
jobs for master and ussuri to be non-voting due to bug [4].

This patch also switches neutron-tempest-plugin-designate-scenario
to be non-voting due to the bug [5]

[1] https://bugs.launchpad.net/neutron/+bug/1889779
[2] https://bugs.launchpad.net/neutron/+bug/1896735
[3] https://bugs.launchpad.net/neutron/+bug/1897326
[4] https://bugs.launchpad.net/networking-bagpipe/+bug/1897408
[5] https://bugs.launchpad.net/neutron/+bug/1891309

Related-Bug: #1896735

Change-Id: I9252b6a8786c43524ba0ebaa59b480ef8e489ff1