Use Cirros 0.5.1 in scenario jobs for all branches
Cirros 0.5.1 has got improvements for getting data from the
metadata service, see [1] for details.
It is used by default in Devstack since Victoria but this patch
defines to download this version of Cirros for all
neutron-tempest-plugin jobs so it will be used also in our
jobs for stable branches.
That should improve our tests reliability as we should see
less Authentication Error while SSHing to the guet vm.

It also should cause any problems as our tests don't rely really
on any specific features of the older Cirros.

This patch additionally updates also advanced image to be
Ubuntu 18.04 where it wasn't changed yet.

[1] https://github.com/cirros-dev/cirros/pull/11/commits/e40bcd2964aa496a9d03e1aaf95cf7a86938f129

Change-Id: I18107d7169b164dfddd1cffb71dffccf2a3ba206