Use different job definition for linuxbridge jobs runs on Xenial
It seems that on Xenial nodes ansible still wants to include
"legacy_ebtables" role which is only available in Neutron repo since
stable/Victoria release.
So lets try to use different job definition, not inherited from
linuxbridge job for master branch for this job on Xenial nodes.

This patch also switches back to use Cirros 0.3.5 (default in devstack)
in stable/queens branch. It was switched to Cirros 0.5.1 in patch [1]
but it seems that this new version of Cirros always has kernel
panic when running on host with Ubuntu Xenial.

[1] https://review.opendev.org/#/c/755112/

Change-Id: I1ceb3a97d29600d8837611d714ff7df70c35fbd7
Closes-Bug: #1899917