Move fwaas jobs to the experimental queue
We have neutron-fwaas related jobs only for the stable/train and
stable/ussuri branches as neutron-fwaas was deprecated from stadium
after ussuri.
In both those cases those jobs are non-voting and broken - both are
failing 100% of the times now.
Based on that there is no point to waste infra resources and run those
jobs on every patch. Lets move them to the experimental queue so it can
be run on demand.

Change-Id: I4ea122e7ca1e1c0b5c00bd3397740510540a8ff9