Remove stein jobs from the project's queues
As Stein is now in EM phase we don't want to run it's jobs for every
patch anymore.
So this patch removes those jobs from the check and gate queue of the
neutron-tempest-plugin repo.
It also freeze to use always "stein-em" tag of the
neutron-tempest-plugin in those jobs as this is last known, good
release which works with Neutron stable/stein.

Depends-On: https://review.opendev.org/c/openstack/releases/+/768553
Change-Id: Ie57b009430bb92f1dabea48399af9faaa7e8987b