Fix multicast scenario test
This patch changes multicast scenario test to run tcpdump on the
"unregistered" host always with specific interface, so that interface
is switched to the promiscuous mode.
It also adds option "-c 1" to the same tcpdump to be sure that it is killed
after it captures one packet.

In OVN job on Ussuri OVN 20.03 is used and IGMP Snooping requires a
least OVN 20.12 so in that version snooping is disabled.

This also reverts commit 193f1714bf4b2cf9bcdee89c0a646cab84544f10 which
marked that multicast scenario test as unstable.

Closes-Bug: #1850288

Change-Id: Ia3c84134a3e0543687ef715025d70f42e4b6826c