Limit execution of the neutron CI jobs on some irrelevant file
This patch tries to make execution of the CI jobs in the Neutron
queue to be more smart. There is no need to e.g. run
neutron-tempest-plugin-api job when only changed files in patch
are in e.g. neutron.agent module.
Or there is no need to run openvswitch scenario jobs if changes
are in e.g. linuxbridge agent only.

Change-Id: I59acf91e4ce396c35d9939ea4887dde3aef55474