Unset new rolevar for run-tempest role for rocky jobs
It was done switched to use new rolevar for newer branches in [1]
some time ago but it seems that as rocky and queens jobs are inheriting
from the jobs defined for master branch, they have both regex variables
set and that is causing problems when running CI jobs for older
branches.
It is like that because it seems that roles defined in Tempest repo
(run-tempest) is always used from master branch, even if tempest is
pinned in devstack to be e.g. 23.0.0 (for Rocky).
Because of that, as for example our job
neutron-tempest-plugin-scenario-openvswitch-iptables_hybrid-rocky
inherits from the neutron-tempest-plugin-scenario-openvswitch-iptables_hybrid
it happened that it had defined both "tempest_black_regex" and
"tempest_exclude_regex" so both tasks:

run-tempest : Build exclude regex (old param)
run-tempest : Build exclude regex (new param)

were skipped and there was no list with tests to skip passed to tempest
really.

To avoid that, we need to define jobs for rocky branch "from
scratch" and not inherit from the jobs defined in zuul.d/master.yaml
as such jobs for master branch is using tempest_exclude_regex variable
and there is no way to unset it in the rocky job.

[1] https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/775257

Change-Id: Iafa52d60050c12ea10ab1bfb4d6147f886303ea5
Closes-Bug: #1924315