Wait for HA router to be active on at least one agent
Due to related bug it may happend that HA router will become active on
one of the L3 agents after few minutes since it was created. And during
that time when it's in "standby" mode, VM can be spawned, it will try
to get metadata but haproxy in the router namespace is still not running
so there will be no metadata available.
That will end up with SSH authentication error as there will be no
proper SSH key configured in the VM.

Originally this patch was intended just as a workaround for the related
bug but I think it's worth to keep it even if that bug will be solved on
the Neutron's side as it may give us more clear information about the
real problem in the future if something similar will happen and HA
router will not be active on any of the L3 agents.

Related-Bug: #1923633
Change-Id: I8c8b7c11c63ffeee4f776695f32ae686793299b3